--- backend/api/editor/routes.ts
@@ -1,11 +1,0 @@
-import { Router } from "express";
-
-const router = Router();
-
-router.post("/save", (req, res) => {
-  console.log(req.body);
-
-  res.send(req.body);
-});
-
-export default router;
