+++ prisma/schema.prisma
@@ -1,0 +1,66 @@
+// This is your Prisma schema file,
+// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+// Looking for ways to speed up your queries, or scale easily with your serverless or edge functions?
+// Try Prisma Accelerate: https://pris.ly/cli/accelerate-init
+
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+model User {
+  id    String  @id @default(uuid())
+  name  String?
+  email String
+  age   Int?
+  posts Post[]
+
+  comments Comment[]
+}
+
+model Category {
+  id    String @id @default(uuid())
+  name  String
+  posts Post[]
+}
+
+enum Page {
+  HOME
+  FRAMES
+  FUNCTIONS
+  ABOUT
+}
+
+model Post {
+  id         String     @id @default(uuid())
+  title      String
+  rating     Float?
+  createdAt  DateTime
+  updatedAt  DateTime
+  author     User       @relation(fields: [authorId], references: [id])
+  authorId   String
+  page       Page?
+  categories Category[]
+  content    String
+
+  Comment Comment[]
+}
+
+model Comment {
+  id               String   @id @default(uuid())
+  author           User?    @relation(fields: [authorId], references: [id])
+  authorId         String?
+  createdAt        DateTime
+  updatedAt        DateTime
+  content          String
+  post             Post     @relation(fields: [postId], references: [id])
+  postId           String
+  replyTo          Comment? @relation("ReplyTo", fields: [replyToCommentId], references: [id])
+  replyToCommentId String
+  Comment Comment[] @relation("ReplyTo")
+}
