+++ src/prisma/client.ts
@@ -1,0 +1,6 @@
+import { PrismaClient } from "@prisma/client";
+
+const prismaClient = new PrismaClient();
+
+export default prismaClient;
+
