+++ src/serverEntry.tsx
@@ -1,7 +1,7 @@
 import React, { StrictMode } from "react";
 import { StaticRouter } from "react-router-dom";
 import { renderToString } from "react-dom/server";
-import App from "../frontend/App";
+import App from "./App";
 
 export const render = (url: string, initialState: AppProps) => {
   return renderToString(
