+++ vite.be.config.ts
@@ -5,14 +5,14 @@ import { defineConfig } from "vite";
 export default defineConfig({
   css: commonConfig.css,
   build: {
-    ssr: "backend/index.ts",
+    ssr: "src/server.ts",
     outDir: path.resolve(__dirname, "dist-server"),
     target: "node23",
     rollupOptions: {
       external: ["express", "fs", "path", "dotenv", "vite", "bun"],
-      input: path.resolve(__dirname, "backend/index.ts"),
+      input: path.resolve(__dirname, "src/server.ts"),
       output: {
-        entryFileNames: "index.js",
+        entryFileNames: "server.js",
         format: "esm", // Use ESM for consistency with modern runtimes
       },
     },
