+++ vite.fe.config.ts
@@ -85,7 +85,7 @@ export default defineConfig(() => {
         ? {
             // In production we have two entry points: one for the client and one for the service worker
             input: {
-              client: path.resolve(__dirname, "frontend/client.tsx"),
+              client: path.resolve(__dirname, "src/client.tsx"),
               "service-worker": path.resolve(
                 __dirname,
                 "web-worker/service-worker.ts",
